project/requirements.txt
fastapi==0.60.1
pandas==1.1.0
plotly==4.9.0
uvicorn==0.11.8
python-dotenv==0.14.0
psycopg2-binary==2.8.6
imbalanced-learn==0.4.3
imblearn==0.0
category-encoders==2.2.2
scikit-learn==0.22.2.post1
